Setup & Auth updates

+++ api/setup/setup.php
@@ -1,9 +1,14 @@
 <?php
+/**
+ * This file deals with setup: if the request method is a 'GET' it will
+ * check whether setup has been already performed; if not, it will set
+ * up GameCourse.
+ */
 
 error_reporting(E_ALL);
 ini_set('display_errors', '1');
 
-use Api\API;
+use API\API;
 use GameCourse\Core\Core;
 use GameCourse\Course\Course;
 use GameCourse\Module\Module;
@@ -13,39 +18,47 @@ require __DIR__ . "/../inc/bootstrap.php";
 
 Core::denyCLI();
 
-// Setup already done; Delete file setup.done to allow
-if (!Core::requireSetup(false))
-    API::error('GameCourse setup already done', 405);
+if ($_SERVER['REQUEST_METHOD'] == "GET") {  // check setup
+    $needsSetup = Core::requireSetup(false);
+    echo json_encode(['isSetupDone' => !$needsSetup]);
 
-if (array_key_exists('course-name', $_POST) && array_key_exists('teacher-id', $_POST)) {
-    $courseName = $_POST['course-name'];
-    $courseColor = $_POST['course-color'];
-    $teacherId = $_POST['teacher-id'];
-    $teacherUsername = $_POST['teacher-username'];
+} else {  // do setup
+    // Setup already done; Delete file setup.done to allow
+    if (!Core::requireSetup(false))
+        API::error('GameCourse setup already done.', 405);
 
-    Core::resetGameCourse();
+    if (array_key_exists('course-name', $_POST) && array_key_exists('course-color', $_POST) &&
+        array_key_exists('teacher-id', $_POST) && array_key_exists('teacher-username', $_POST)) {
 
-    // Init database
-    $sql = file_get_contents(ROOT_PATH . "setup/setup.sql");
-    Core::database()->executeQuery($sql);
+        $courseName = $_POST['course-name'];
+        $courseColor = $_POST['course-color'];
+        $teacherId = $_POST['teacher-id'];
+        $teacherUsername = $_POST['teacher-username'];
 
-    // Create user in the system
-    $user = User::addUser("Teacher", $teacherUsername, "fenix", null, $teacherId, null,
-        null, true, true);
-    Core::setLoggedUser($user);
+        Core::resetGameCourse();
 
-    // Register modules available
-    Module::setupModules();
+        // Init database
+        $sql = file_get_contents(ROOT_PATH . "setup/setup.sql");
+        Core::database()->executeQuery($sql);
 
-    // Create course
-    // NOTE: user is automatically added as a teacher of the course
-    $course = Course::addCourse($courseName, null, null, $courseColor, null, null,
-        true, true);
+        // Create user in the system
+        $user = User::addUser("Teacher", $teacherUsername, "fenix", null, $teacherId, null,
+            null, true, true);
+        Core::setLoggedUser($user);
 
-    file_put_contents("./setup.done", "");
-    echo json_encode(['setup' => true]);
-    exit;
+        // Register modules available
+        Module::setupModules();
+
+        // Create course
+        // NOTE: user is automatically added as a teacher of the course
+        $course = Course::addCourse($courseName, null, null, $courseColor, null, null,
+            true, true);
+
+        file_put_contents(ROOT_PATH . "setup/setup.done", "");
+        echo json_encode(['setup' => true]);
+        exit;
+    }
+
+    API::error('Some information is missing. Please fill-in all the fields.');
 }
-
-echo json_encode(['setup' => false]);
